refactor config docgen and add the rest of the config types
