+++ docs/config/console.md
@@ -1,0 +1,25 @@
+# console
+
+* `console`  
+  
+  * `account_namespace`  
+    Set account namespace in console context if set true.  
+    All accounts will be available in console context if set false  
+    
+    
+  * `contract_namespace`  
+    Set contract namespace in console context if set true.  
+    All contracts will be available in console context if set false  
+    
+    
+  
+
+---
+
+## Default Config
+
+```toml
+[console]
+account_namespace = true
+contract_namespace = true
+```
