+++ docs/config/wasm.md
@@ -1,0 +1,25 @@
+# wasm
+
+* `wasm`  
+  
+  * `contract_dir`  
+    
+    
+  * `template_repo`  
+    
+    
+  * `optimizer_version`  
+    
+    
+  
+
+---
+
+## Default Config
+
+```toml
+[wasm]
+contract_dir = 'contracts'
+template_repo = 'https://github.com/InterWasm/cw-template.git'
+optimizer_version = '0.12.6'
+```
