+++ packages/gen_docs/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod cmark;
+pub mod command;
+pub mod config;
+pub mod document;
